updated return data without password

+++ backend/src/models/User.js
@@ -2,7 +2,9 @@ import { pool } from "../config/db.js";
 
 export class User {
   static async getUsers() {
-    const data = await pool.query(`SELECT * FROM users`);
+    const data = await pool.query(
+      `SELECT user_id, name, surname, patronymic, login, id_role  * FROM users`,
+    );
     return data.rows;
   }
 
@@ -20,7 +22,7 @@ export class User {
         patronymic,
         login,
         password,
-        id_role) VALUES($1, $2, $3, $4, $5, $6) RETURNING *`,
+        id_role) VALUES($1, $2, $3, $4, $5, $6) RETURNING user_id, name, surname, patronymic, login, id_role`,
       [name, surname, patronymic, login, password, id_role],
     );
 
@@ -31,7 +33,7 @@ export class User {
 
   static async deleteUser(id_user) {
     const data = await pool.query(
-      `DELETE FROM users WHERE id_user=$1 RETURNING *`,
+      `DELETE FROM users WHERE id_user=$1 RETURNING user_id, name, surname, patronymic, login, id_role`,
       [id_user],
     );
 
@@ -41,9 +43,10 @@ export class User {
   }
 
   static async getUser(id_user) {
-    const data = await pool.query(`SELECT * FROM users WHERE id_user=$1`, [
-      id_user,
-    ]);
+    const data = await pool.query(
+      `SELECT user_id, name, surname, patronymic, login, id_role FROM users WHERE id_user=$1`,
+      [id_user],
+    );
 
     if (!data.rows.length) throw { status: 404, message: "User Not Found" };
